backend/requirements.txt
bcrypt
python-dotenv
PyJWT
